--- activity_menu.xml
@@ -1,109 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/menu_frame"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:layout_constraintTop_toTopOf="parent"
-    app:layout_constraintBottom_toTopOf="@+id/menu_scrollview">
-
-        <ImageButton
-        android:contentDescription="profile button"
-        android:id="@+id/menu_imgbutton_profile"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:scaleX="0.8" android:scaleY="0.8"
-        android:src="@drawable/profile_small"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/menu_textview_username"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Username"
-        android:textSize="20sp"
-        app:layout_constraintLeft_toRightOf="@+id/menu_imgbutton_profile"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
-    <ScrollView
-        android:id="@+id/menu_scrollview"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/menu_frame"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_marginRight="40dp"
-        android:layout_marginLeft="40dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_scalecalculator"
-                android:text="Scale Calculator"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_noteid"
-                android:text="Note identification"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_chordid"
-                android:text="Chord identification"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_intervalid"
-                android:text="Interval identification"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_notesound"
-                android:text="Note Sound"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-            <Button
-                android:padding="10dp"
-                android:id="@+id/menu_button_intervalsound"
-                android:text="Interval Sound"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:layout_marginBottom="6dp"
-                android:layout_marginTop="6dp"/>
-        </LinearLayout>
-    </ScrollView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
